handle get location
