+++ app/src/main/java/com/example/mba0229p/da_nang_travel/data/model/event/LocationEvent.kt
@@ -1,0 +1,2 @@
+package com.example.mba0229p.da_nang_travel.data.model.event
+
